Ajout de la méthode Update

+++ src/main/java/com/example/projet3chatop/dto/RentalDto.java
@@ -22,8 +22,8 @@ public class RentalDto {
     @JoinColumn(name = "owner_id", referencedColumnName = "id")
     private User ownerId;
     private String name;
-    private int surface;
-    private double price;
+    private float surface;
+    private float price;
     private String description;
